refactor-03: Error Handling Consolidation
